+++ drivers/mtd/nand/raw/sp_bch_q645.c
@@ -125,7 +125,7 @@ static int sp_bch_reset(struct sp_spinand_info *info)
 		return ret;
 	}
 
-	writel(~(IER_FAIL|IER_DONE), &regs->ier);
+	writel((uint32_t) (~(IER_FAIL|IER_DONE)), &regs->ier);
 	writel(SR_DONE|SR_FAIL, &regs->sr);
 	writel(ISR_BCH, &regs->isr);
 
@@ -338,8 +338,8 @@ int sp_bch_encode(struct mtd_info *mtd, void *buf, void *ecc)
 	flush_dcache_range((ulong) ecc, (ulong) ecc + mtd->oobsize);
 
 	writel(SRR_RESET, &regs->srr);
-	writel((uint32_t) buf, &regs->buf);
-	writel((uint32_t) ecc, &regs->ecc);
+	writel((uint32_t)(uint64_t) buf, &regs->buf);
+	writel((uint32_t)(uint64_t) ecc, &regs->ecc);
 
 	writel(CR0_START | CR0_ENCODE | info->cr0, &regs->cr0);
 	ret = sp_bch_wait(info);
@@ -363,8 +363,8 @@ int sp_bch_encode_1024x60(void *buf, void *ecc)
 	flush_dcache_range((ulong) ecc, (ulong) ecc + 128);
 
 	writel(SRR_RESET, &regs->srr);
-	writel((uint32_t) buf, &regs->buf);
-	writel((uint32_t) ecc, &regs->ecc);
+	writel((uint32_t)(uint64_t) buf, &regs->buf);
+	writel((uint32_t)(uint64_t) ecc, &regs->ecc);
 
 	writel(CR0_BMODE(5)|CR0_START | CR0_ENCODE | CR0_CMODE_1024x60, &regs->cr0);
 
@@ -389,8 +389,8 @@ int sp_bch_decode_1024x60(void *buf, void *ecc)
 	flush_dcache_range((ulong) ecc, (ulong) ecc + 128);
 
 	writel(SRR_RESET, &regs->srr);
-	writel((uint32_t) buf, &regs->buf);
-	writel((uint32_t) ecc, &regs->ecc);
+	writel((uint32_t)(uint64_t) buf, &regs->buf);
+	writel((uint32_t)(uint64_t) ecc, &regs->ecc);
 
 	writel(CR0_BMODE(5)|CR0_START | CR0_DECODE | CR0_CMODE_1024x60, &regs->cr0);
 
@@ -431,8 +431,8 @@ int sp_bch_decode(struct mtd_info *mtd, void *buf, void *ecc)
 	flush_dcache_range((ulong) buf, (ulong) buf + mtd->writesize);
 	flush_dcache_range((ulong) ecc, (ulong) ecc + mtd->oobsize);
 
-	writel((uint32_t) buf, &regs->buf);
-	writel((uint32_t) ecc, &regs->ecc);
+	writel((uint32_t)(uint64_t) buf, &regs->buf);
+	writel((uint32_t)(uint64_t) ecc, &regs->ecc);
 	writel(CR0_START | CR0_DECODE | info->cr0, &regs->cr0);
 
 	ret = sp_bch_wait(info);
@@ -464,8 +464,8 @@ int sp_autobch_config(struct mtd_info *mtd, void *buf, void *ecc, int enc, int d
 	int value;
 
 	writel(SRR_RESET, &regs->srr);
-	writel((uint32_t) buf, &regs->buf);
-	writel((uint32_t) ecc, &regs->ecc);
+	writel((uint32_t)(uint64_t) buf, &regs->buf);
+	writel((uint32_t)(uint64_t) ecc, &regs->ecc);
 
 	value = info->cr0
 		| (enc ? CR0_ENCODE : CR0_DECODE)
